13-MCP-with-ADK/requirements.txt
google-adk==1.4.1
mcp==1.9.1